moved ssh, cli, and kubernetes under root subfolder
