+++ cmd/root/collection/sudo.go
@@ -16,7 +16,7 @@
 package collection
 
 import (
-	"github.com/dremio/dremio-diagnostic-collector/cli"
+	"github.com/dremio/dremio-diagnostic-collector/cmd/root/cli"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
 )
