+++ cmd/root/kubernetes/kubernetes_test.go
