+++ cmd/root/ssh/ssh.go
@@ -19,7 +19,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/dremio/dremio-diagnostic-collector/cli"
+	"github.com/dremio/dremio-diagnostic-collector/cmd/root/cli"
 )
 
 func NewCmdSSHActions(sshKey, sshUser string) *CmdSSHActions {
