+++ pkg/tests/mocks.go
@@ -15,7 +15,7 @@
 // package tests provides helper functions and mocks for running tests
 package tests
 
-import "github.com/dremio/dremio-diagnostic-collector/cli"
+import "github.com/dremio/dremio-diagnostic-collector/cmd/root/cli"
 
 type MockCli struct {
 	Calls          [][]string
